Add Department page

+++ src/main/java/com/moldavets/finalproject/service/EmployeeService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface EmployeeService {
 
     List<Employee> getAll();
-    Employee getById(Long id);
+    Employee getById(int id);
     void save(Employee employee);
-    void deleteById(Long id);
+    void deleteById(int id);
 }
